% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gk_arrange.R
\name{gk_arrange}
\alias{gk_arrange}
\title{Veri Setinin Duzenlenmesi}
\usage{
gk_arrange(
  data,
  proj_yascins,
  proj_il,
  proj_nutskirkent,
  proj_hhbkirkent,
  cinsiyet = "CINSIYET",
  birimno = "BIRIMNO",
  yas = "YAS",
  hhbsay = 9,
  hhbkirkent = TRUE
)
}
\arguments{
\item{data}{Duzenlenmemis HIA veri seti. \cr
Bu veri seti uzerinde \strong{yasgrubu, cinsiyet, hhb, IBBS_2} gibi
degiskenlerde duzenlemeler yapilarak dummy degiskenler olusturulmaktadir. \cr
Ayrica \strong{"gk_caltrim"} fonksiyonunda, kalibrasyon icin kullanilmak uzere projekte populasyonlari da veri setine eklenmektedir. \cr
\cr
\emph{NOT:} Veri setinde yer alacak degiskenlerin standart bir formatta olmasi gerekmektedir. \cr
\cr
Ornegin; \cr
\itemize{
\item ILKAYITNO: veri tipi character ; 01, 02, 03 seklinde
\item IBBS_1: veri tipi character ; TR1, TR2 seklinde
\item IBBS_2: veri tipi character ; TR10, TR21 seklinde
\item KIR_KENT: veri tipi numeric ; 1, 2 seklinde
\item CINSIYET: veri tipi numeric ; 1, 2 seklinde
\item FAKTOR: veri tipi numeric
\item HHB: veri tipi numeric ; 1, 2 ,3 seklinde
\item duzfakto: veri tipi numeric
\item fk_fkiml: veri tipi numeric ; 1, 2, 3 seklinde
}}

\item{proj_yascins}{Yas - Cinsiyet projekte oranlari \emph{(tek satir halinde)}}

\item{proj_il}{IL projekte oranlari \emph{(tek satir halinde)}}

\item{proj_nutskirkent}{NUTS2 - Kirkent projekte oranlari \emph{(tek satir halinde)}}

\item{proj_hhbkirkent}{HHB - Kirkent ADNKS oranlari \emph{(Kirkent ayrimi olacaksa HHB*2,
kirkent ayrimi olmayacaksa HHB grup sayisi kadar satir olacak)} \cr}

\item{cinsiyet}{Cinsiyet degiskeni icin kullanilacak sutun adi \emph{(veri tipi numeric: 1, 2 seklinde)}}

\item{birimno}{Birimno sutun adi \emph{(veri tipi numeric: 325069, 325070 seklinde)}}

\item{yas}{Yas sutun adi \emph{(veri tipi numeric: 26, 35 seklinde)}}

\item{hhbsay}{Hanehalki buyuklugu degiskeninin kac grup olacagini belirtmektedir. Ornegin "hhbsay = 4"
degeri kullanildiginda kalibrasyon islemleri 1, 2, 3 ve 4+ hanehalki buyuklugu gruplarinda yapilacaktir.}

\item{hhbkirkent}{HHB * KIRKENT crossunda kalibrasyon yapilmasi isteniyorsa TRUE,
sadece HHB bazinda kalibrasyon yapilmasi isteniyorsa FALSE degerini almaktadir.}
}
\value{
\strong{"gk_caltrim"} fonksiyonuna girdi olacak, duzenlenmis PERS dosyasini cikti olarak verir. \cr
Detayli aciklama icin \link{gk_caltrim} fonksiyonunu inceleyiniz...
}
\description{
Veri Setinin Duzenlenmesi
}
