% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gk_check.R
\name{gk_check}
\alias{gk_check}
\title{Veri Setinin Kontrolu}
\usage{
gk_check(data)
}
\arguments{
\item{data}{Check edilecek veri seti}
}
\value{
Kontrol sonrasi TRUE veya FALSE ciktisini verir
}
\description{
Veri Setinin Kontrolu
}
\examples{
control <- gk_check(mtcars);
if(gk_check(mtcars)) {x <- 999};
}
