% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gklfs.R
\docType{package}
\name{gklfs}
\alias{gklfs}
\title{gklfs: LFS veri setinin agirliklandirmasi ve varyans hesaplamalari}
\description{
Hanehalki Isgucu verilerinin agirliklandirma islemleri yapilmaktadir.
Pakette 4 farkli fonksiyon bulunmaktadir.
gk_arrange, gk_check, gk_caltrim ve gk_variance
}
\section{gklfs fonksiyonlari}{

gk_arrange: Kalibrasyon ve trim islemlerine girecek veri setinin duzenlenmesi,
gk_check: Veri setinde eksik (missing) veya yanlis veri olup olmadiginin kontrol edilmesi,
gk_caltrim: Kullanici tarafindan belirlenen kriterlere gore kalibrasyon ve trimming islemlerinin yapilarak
nihai agirliklarin elde edilmesi,
gk_variance: Kullanici tarafindan girilen modele gore Taylor Linearization yontemiyle varyans, CV ve deff
degerlerinin hesaplanmasi
amaciyla kullanilmaktadir.
}

