% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gk_arrange.R
\name{gk_arrange}
\alias{gk_arrange}
\title{Veri Setinin DuzenlenmesiPPP}
\usage{
gk_arrange(
  data,
  proj_yascins,
  proj_il,
  proj_nutskirkent,
  proj_hhbkirkent,
  cinsiyet = "CINSIYET",
  birimno = "BIRIMNO",
  yas = "YAS",
  hhbsay = 9,
  hhbkirkent = TRUE
)
}
\arguments{
\item{data}{Duzenlenmemis veri seti. Bu veri seti kullanilarak yasgrubu, cinsiyet, hhb, IBBS_2 gibi
degiskenlerde duzenlemeler yapiliyor. Ayrica "gk_caltrim" fonksiyonunda kalibrasyon icin kullanilacak
dummy degiskenler ve projekte populasyonlari da veri setine ekleniyor.}

\item{proj_yascins}{Yas - Cinsiyet projekte oranlari (tek satir halinde)}

\item{proj_il}{IL projekte oranlari (tek satir halinde)}

\item{proj_nutskirkent}{NUTS2 - Kirkent projekte oranlari (tek satir halinde)}

\item{proj_hhbkirkent}{HHB - Kirkent ADNKS oranlari (Kirkent ayrimi olacaksa HHB*2,
kirkent ayrimi olmayacaksa HHB grup sayisi kadar satir olacak)}

\item{cinsiyet}{Cinsiyet sutun adi (veri tipi numeric: 1, 2 seklinde)}

\item{birimno}{Birimno sutun adi (veri tipi numeric: 325069, 325070 seklinde)}

\item{yas}{Yas sutun adi (veri tipi numeric: 26, 35 seklinde)}

\item{hhbsay}{Hanehalki buyuklugunun kac grup olacagi}

\item{hhbkirkent}{HHB * KIRKENT crossunda kalibrasyon yapilmasi isteniyorsa TRUE,
sadece HHB bazinda kalibrasyon yapilmasi isteniyorsa FALSE degerini alir.}
}
\value{
"gk_caltrim" fonksiyonuna girdi olacak duzenlenmis PERS dosyasini olusturur.
}
\description{
Veri Setinin DuzenlenmesiPPP
}
