% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gk_arrange.R
\name{gk_arrange}
\alias{gk_arrange}
\title{Veri Setinin Duzenlenmesi}
\usage{
gk_arrange(
  data,
  proj_yascins,
  proj_il,
  proj_nutskirkent,
  proj_hhbkirkent,
  cinsiyet = "CINSIYET",
  birimno = "BIRIMNO",
  yas = "YAS",
  hhbsay = 9,
  hhbkirkent = TRUE
)
}
\arguments{
\item{data}{Duzenlenmemis HIA veri seti. \cr
Bu veri seti kullanilarak yasgrubu, cinsiyet, hhb, IBBS_2 gibi
degiskenlerde duzenlemeler yapilarak dummy degiskenler olusturulmaktadir. \cr
Ayrica "gk_caltrim" fonksiyonunda, kalibrasyon icin kullanilmak uzere projekte populasyonlari da veri setine eklenmektedir. \cr
NOT: Veri setinde yer alacak degiskenlerin standart bir formatta olmasi gerekmektedir. \cr
Ornegin; \cr
ILKAYITNO: veri tipi character ; 01, 02, 03 seklinde \cr
IBBS_1: veri tipi character ; TR1, TR2 seklinde \cr
IBBS_2: veri tipi character ; TR10, TR21 seklinde \cr
KIR_KENT: veri tipi numeric ; 1, 2 seklinde \cr
CINSIYET: veri tipi numeric ; 1, 2 seklinde \cr
FAKTOR: veri tipi numeric \cr
HHB: veri tipi numeric ; 1, 2 ,3 seklinde \cr
duzfakto: veri tipi numeric \cr
fk_fkiml: veri tipi numeric ; 1, 2, 3 seklinde}

\item{proj_yascins}{Yas - Cinsiyet projekte oranlari (tek satir halinde)}

\item{proj_il}{IL projekte oranlari (tek satir halinde)}

\item{proj_nutskirkent}{NUTS2 - Kirkent projekte oranlari (tek satir halinde)}

\item{proj_hhbkirkent}{HHB - Kirkent ADNKS oranlari (Kirkent ayrimi olacaksa HHB*2,
kirkent ayrimi olmayacaksa HHB grup sayisi kadar satir olacak)}

\item{cinsiyet}{Cinsiyet degiskeni icin kullanilacak sutun adi (veri tipi numeric: 1, 2 seklinde)}

\item{birimno}{Birimno sutun adi (veri tipi numeric: 325069, 325070 seklinde)}

\item{yas}{Yas sutun adi (veri tipi numeric: 26, 35 seklinde)}

\item{hhbsay}{Hanehalki buyuklugu degiskeninin kac grup olacagini belirtmektedir. Ornegin "hhbsay = 4"
degeri kullanildiginda kalibrasyon islemleri 1, 2, 3 ve 4+ hanehalki buyuklugu gruplarinda yapilacaktir.}

\item{hhbkirkent}{HHB * KIRKENT crossunda kalibrasyon yapilmasi isteniyorsa TRUE,
sadece HHB bazinda kalibrasyon yapilmasi isteniyorsa FALSE degerini almaktadir.}
}
\value{
"gk_caltrim" fonksiyonuna girdi olacak duzenlenmis PERS dosyasini cikti olarak verir.
}
\description{
Veri Setinin Duzenlenmesi
}
