% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gk_caltrim.R
\name{gk_caltrim}
\alias{gk_caltrim}
\title{Kalibrasyon & Trim}
\usage{
gk_caltrim(
  data,
  proj_yascins,
  proj_il,
  proj_nutskirkent,
  proj_hhbkirkent,
  calyap = TRUE,
  calsay = 8,
  trimyap = TRUE,
  trimsay = 4,
  hhbsay = 9,
  hhbkirkent = TRUE,
  yascinsiyet = TRUE,
  il = TRUE,
  nutskirkent = TRUE,
  kirnufus = 9e+06,
  kentnufus = 7e+07
)
}
\arguments{
\item{data}{Duzenlenmis, dummyleri olusturulmus, projekte oranlari eklenmis veri seti.
gk_arrange" fonksiyonundan elde edilen veri seti burada kullanilmalidir.}

\item{calyap}{Kalibrasyon yapilacaksa TRUE, yapilmayacaksa FALSE degerini alir.}

\item{calsay}{Kalibrasyon dongu sayisi}

\item{trimyap}{Trim yapilacaksa TRUE, yapilmayacaksa FALSE degerini alir.}

\item{trimsay}{Trim dongu sayisi}

\item{hhbsay}{Hanehalki buyuklugune gore olusturulacak grup sayisi. Ornegin
"hhbsay = 4" girildiginde hanehalki buyuklugu gruplari 1, 2, 3, 4+ seklinde olusacaktir.
gk_arrange" fonksiyonunda da ayni arguman bulunuyor. Ikisi birbirinden farkli
olduğunda dummy uyusmazligi nedeniyle program hata verecektir. O yuzden veri setinin duzenlenmesi (gk_arrange)
ve agirliklandirilmasi (gk_caltrim) asamalarinda ayni degeri almalidir.}

\item{hhbkirkent}{HHB * KIRKENT crossunda kalibrasyon yapilmasi isteniyorsa TRUE,
sadece HHB bazinda kalibrasyon yapilmasi isteniyorsa FALSE degerini alir. "gk_arrange"
fonksiyonunda da ayni arguman bulunuyor. Ikisi birbirinden farkli olduğunda dummy uyusmazligi
nedeniyle program hata verecektir. Ornegin "gk_arrange" fonksiyonu icerisinde "hhbkirkent = TRUE"
olarak tanimlandiginda dummyler "D_hhb_kk_1_2" seklinde olusturulurken "gk_caltrim" fonksiyonu
icerisinde "hhbkirkent = FALSE" olarak tanimlanirsa dummyler "D_hhb_1" seklinde olusacagi icin
program hata verecektir. O yuzden veri setinin duzenlenmesi (gk_arrange)
ve agirliklandirilmasi (gk_caltrim) asamalarinda ayni degeri almalidir.}

\item{yascinsiyet}{YAS * CINSIYET'e gore kalibrasyon yapilmasi isteniyorsa TRUE,
YAS * CINSIYET'e gore kalibrasyon yapilmasi istenmiyorsa FALSE degerini alir.}

\item{il}{IL'e gore kalibrasyon yapilmasi isteniyorsa TRUE, IL'e gore kalibrasyon
yapilmasi istenmiyorsa FALSE degerini alir.}

\item{nutskirkent}{NUTS2 * KIRKENT crossunda kalibrasyon yapilmasi isteniyorsa TRUE,
NUTS2 * KIRKENT crossunda kalibrasyon yapilmasi istenmiyorsa FALSE degerini alir.}

\item{kirnufus}{Projekte KIR toplam nufusu}

\item{kentnufus}{Projekte KENT toplam nufusu}
}
\value{
Kalibrasyon sonrasi elde edilen agirliklarla birlikte temel degiskenlerin yer
aldigi veri setini cikti olarak verir. Bu veri seti uzerinden issizlik, istihdam oranlari
hesaplanabilir.
}
\description{
Kalibrasyon & Trim
}
