% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gk_check.R
\name{gk_check}
\alias{gk_check}
\title{Veri Setinin Kontrolu}
\usage{
gk_check(data)
}
\arguments{
\item{data}{Kontrol edilecek veri seti. Kalibrasyon oncesinde DUZFAKTO, CINSIYET, ILKAYITNO,
NUTS2*KIRKENT ve HHB_KIRKENT degiskenlerinde yanlis veya missing (NA) veri olup olmadigini
kontrol etmektedir.}
}
\value{
Kontrol sonrasinda TRUE veya FALSE ciktisini verir. Ekranda kontrol sonuclarini mesajla gosterir.
}
\description{
Veri Setinin Kontrolu
}
\examples{
control <- gk_check(mtcars);
if(gk_check(mtcars)) {x <- 999};
}
